requirements.txt
asgiref==3.11.1
Django==5.2.11
django-ranged-response==0.2.0
django-simple-captcha==0.6.3
gunicorn==25.1.0
packaging==26.0
pillow==12.1.1
sqlparse==0.5.5
tzdata==2025.3
whitenoise==6.12.0